+++ tests/Types/UomTest.php
@@ -13,11 +13,11 @@ class UomTest extends TestCase
     public function testFields(): void
     {
         $uom = new Uom([
-            Uom::TYPE     => 'SEAT',
             Uom::QUANTITY => 1,
+            Uom::TYPE     => 'SEAT',
         ]);
 
-        self::assertSame(1, $uom->getQuantity());
         self::assertSame('SEAT', $uom->getType());
+        self::assertSame(1, $uom->getQuantity());
     }
 }